requirements.txt
# docs stuff
mkdocs>=1.5.0
mkdocs-terminal>=4.0.0

opencv-python>=4.8.0
roboflow>=1.1.0
inference>=0.9.0
supervision>=0.16.0
easyocr>=1.7.0
onnxruntime
Pillow
PyQt6
PyYAML
scipy

mss

# ML / Decision Transformer
torch>=2.0.0
torchvision
ultralytics
numpy
scikit-learn
joblib
tqdm

python-dotenv
